requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-login==0.6.3
flask-sqlalchemy==3.1.1
ultralytics==8.1.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
werkzeug==3.0.1
eventlet==0.35.1
python-dotenv==1.0.1
PyYAML==6.0.1
